feat: rename ipfs to beacon
