+++ src/services/beacon.ts
